# be/openapi.yaml
openapi: 3.0.3
info:
  title: Memory Game API
  description: Backend API for the Memory Game (cards and scores).
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development

paths:
  /players:
    get:
      summary: Check if player name is unique
      description: Returns whether the given player name is available (unique) or already taken, based on the Scores table.
      operationId: getPlayersCheck
      parameters:
        - name: name
          in: query
          required: true
          description: Player name to check
          schema:
            type: string
      responses:
        '200':
          description: Uniqueness result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerUnique'
        '400':
          description: Missing or empty name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cards:
    get:
      summary: Get cards by theme
      description: Returns a list of cards for the given theme, limited by pair count (for memory game pairs).
      operationId: getCards
      parameters:
        - name: theme
          in: query
          description: Card theme. Must be one of dogs, animals, flags, food, plants.
          required: false
          schema:
            type: string
            enum: [dogs, animals, flags, food, plants]
            default: dogs
        - name: pairCount
          in: query
          description: Number of card pairs to return (1–100).
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 8
      responses:
        '200':
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
        '400':
          description: Invalid theme (must be one of dogs, animals, flags, food, plants)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  allowed: { type: array, items: { type: string } }

  /score:
    post:
      summary: Store a score
      description: Saves a player score. Player name must be unique (not already present in Scores). Date/time is set by the API.
      operationId: postScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreInput'
      responses:
        '204':
          description: Score stored successfully (no content)
        '400':
          description: Invalid request body or player name already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a player's score
      description: Updates the score for the given player name. Date/time is set by the API. Returns 404 if player is not in Scores.
      operationId: putScore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreInput'
      responses:
        '204':
          description: Score updated successfully (no content)
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /scores:
    get:
      summary: Get score leaderboard
      description: Returns scores sorted by score descending, 25 per page, with global rank.
      operationId: getScores
      parameters:
        - name: page
          in: query
          description: Page number (1-based).
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: Page of scores with rank and dateTime
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreEntry'

components:
  schemas:
    PlayerUnique:
      type: object
      required: [unique]
      properties:
        unique:
          type: boolean
          description: true if name is available, false if already taken
          example: true

    Card:
      type: object
      required: [id, theme, name, url]
      properties:
        id:
          type: integer
          example: 1
        theme:
          type: string
          example: dogs
        name:
          type: string
          example: terrier-sealyham
        url:
          type: string
          format: uri
          example: https://images.dog.ceo/breeds/terrier-sealyham/n02095889_975.jpg

    ScoreInput:
      type: object
      required: [playerName, score]
      properties:
        playerName:
          type: string
          example: name
        score:
          type: number
          example: 232

    ScoreEntry:
      type: object
      required: [playerName, score, rank, dateTime]
      properties:
        playerName:
          type: string
          example: name3
        score:
          type: number
          example: 8329
        rank:
          type: integer
          description: Global 1-based rank (e.g. 1–25 on page 1)
          example: 1
        dateTime:
          type: string
          format: date-time
          description: When the score was stored (ISO 8601)
          example: "2025-03-15T13:45:30.000Z"

    Error:
      type: object
      properties:
        error:
          type: string
          example: "playerName (string) and score (number) required"